Fix UI show image again

+++ libs/core/src/lib/chat/contexts/ChatContext.tsx
@@ -313,23 +313,24 @@ const ChatContextProvider: React.FC<ChatContextProviderProps> = ({ children }) =
 					mess.isCurrentChannel = message.channel_id === idToCompare;
 				}
 
-				const attachmentList: AttachmentEntity[] = (message.attachments || [])?.map((attachment) => {
-					const dateTime = new Date();
-					return {
-						...attachment,
-						id: attachment.url as string,
-						message_id: message?.message_id,
-						create_time: dateTime.toISOString(),
-						uploader: message?.sender_id
-					};
-				});
+				const attachmentList: AttachmentEntity[] =
+					message.attachments && message.attachments.length > 0
+						? message.attachments.map((attachment) => {
+								const dateTime = new Date();
+								return {
+									...attachment,
+									id: attachment.url as string,
+									message_id: message?.message_id,
+									create_time: dateTime.toISOString(),
+									uploader: message?.sender_id
+								};
+							})
+						: [];
 
-				if (attachmentList?.length) {
-					if (message?.code === TypeMessage.Chat) {
-						dispatch(attachmentActions.addAttachments({ listAttachments: attachmentList, channelId: message.channel_id }));
-					} else if (message?.code === TypeMessage.ChatRemove) {
-						dispatch(attachmentActions.removeAttachments({ messageId: message?.message_id as string, channelId: message.channel_id }));
-					}
+				if (attachmentList?.length && message?.code === TypeMessage.Chat) {
+					dispatch(attachmentActions.addAttachments({ listAttachments: attachmentList, channelId: message.channel_id }));
+				} else if (message?.code === TypeMessage.ChatRemove && message?.attachments) {
+					dispatch(attachmentActions.removeAttachments({ messageId: message?.message_id as string, channelId: message.channel_id }));
 				}
 
 				dispatch(messagesActions.addNewMessage(mess));
